# swagger.yaml
openapi: 3.0.0
info:
  title: Invoice backend
  description:  Invoice API docs
  version: 1.0.0
paths:
  /candidate:
    post:
      tags:
      - "candidate"
      description: "Create candidates"
      requestBody:
       description: Schema of candidate
       required: true
       content:
        application/json:
         schema:
          $ref: '#/components/schemas/Candidate'
      responses:
        "201":
          description: Success Request
          content:
           application/json:
            schema:
              type: "object"
              properties:
               status:
                type: "object"
                properties:
                 success:
                  type: "boolean"
                  example: true
                 code:
                  type: "integer"
                  example: 201
                 message:
                   type: "string"
                   example: request successful candidate created succesfully
        "400":
          description: Bad Request
          content:
           application/json:
            schema:
              type: "object"
              properties:
               status:
                type: "object"
                properties:
                 success:
                  type: "boolean"
                  example: false
                 code:
                  type: "integer"
                  example: 400
                 message:
                   type: "string"
                   example: Please enter ____
        "500":
          description: Failed request
          content:
           application/json:
            schema:
              type: "object"
              properties:
               status:
                type: "object"
                properties:
                 success:
                  type: "boolean"
                  example: false
                 code:
                  type: "integer"
                  example: 500
                 message:
                   type: "string"
                   example: error message
                        
    get:
      tags:
      - "candidate"
      description: "Get All Candidates"
      responses:
        "200":
          description: Success Request
          content:
           application/json:
            schema:
              type: "object"
              properties:
               status:
                type: "object"
                properties:
                 success:
                  type: "boolean"
                  example: true
                 code:
                  type: "integer"
                  example: 200
                 message:
                   type: "string"
                   example: request successful 
               data:
                type: "array"
                items:
                 type: "object"
                 properties:
                  id:
                   type: "string"
                   example: 61f9446265ade063b02174c8
                  username:
                   type: "string"
                   example: "Harshith"
                  email:
                   type: "string"
                   example: "example.p@gmail.com"
                  date_of_birth:
                   type: "string"
                   example: "2/1/1998"
                  assigned_to:
                   type: "string"
                   example: "details"
                  date_of_Joining:
                   type: "string"
                   example: "1/7/2021"
                  payment_terms:
                   type: "string"
                   example: "Terms"
                  role:
                   type: "string"
                   example: "Candidate"
                  isActive:
                   type: "boolean"
                   example: true
        "404":
          description: Failed request
          content:
           application/json:
            schema:
              type: "object"
              properties:
               status:
                type: "object"
                properties:
                 success:
                  type: "boolean"
                  example: false
                 code:
                  type: "integer"
                  example: 404
                 message:
                   type: "string"
                   example: No candidates 
        "500":
          description: Failed request
          content:
           application/json:
            schema:
              type: "object"
              properties:
               status:
                type: "object"
                properties:
                 success:
                  type: "boolean"
                  example: false
                 code:
                  type: "integer"
                  example: 500
                 message:
                   type: "string"
                   example: error message
          
  /candidate/{candidate_id}:
    put:
      tags:
      - "candidate"
      description: "Updates a Candidate"
      parameters:
      - name: candidate_id
        required: true
        in: path
        schema:
          type: integer
      requestBody:
       description: Schema of PUT Request
       required: true
       content:
        application/json:
         schema:
          $ref: '#/components/schemas/Candidate'
      responses:
        "200":
          description: Success Request
          content:
           application/json:
            schema:
              type: "object"
              properties:
               status:
                type: "object"
                properties:
                 success:
                  type: "boolean"
                  example: true
                 code:
                  type: "integer"
                  example: 204
                 message:
                   type: "string"
                   example: request successful candidate updated successfully
        "400":
          description: Bad request
          content:
           application/json:
            schema:
              type: "object"
              properties:
               status:
                type: "object"
                properties:
                 success:
                  type: "boolean"
                  example: false
                 code:
                  type: "integer"
                  example: 400
                 message:
                   type: "string"
                   example: Please enter ____
        "404":
          description: Failed request
          content:
           application/json:
            schema:
              type: "object"
              properties:
               status:
                type: "object"
                properties:
                 success:
                  type: "boolean"
                  example: false
                 code:
                  type: "integer"
                  example: 404
                 message:
                   type: "string"
                   example: candidate not found
        "500":
          description: Failed request
          content:
           application/json:
            schema:
              type: "object"
              properties:
               status:
                type: "object"
                properties:
                 success:
                  type: "boolean"
                  example: false
                 code:
                  type: "integer"
                  example: 500
                 message:
                   type: "string"
                   example: error message
    get:
      tags:
      - "candidate"
      description: "Get Candidate"
      parameters:
      - name: candidate_id
        required: true
        in: path
        schema:
          type: integer
      responses:
        "200":
          description: Success Request
          content:
           application/json:
            schema:
              type: "object"
              properties:
               status:
                type: "object"
                properties:
                 success:
                  type: "boolean"
                  example: true
                 code:
                  type: "integer"
                  example: 200
                 message:
                   type: "string"
                   example: request successful
               data:
                type: "object"
                properties:
                  id:
                   type: "string"
                   example: 61f9446265ade063b02174c8
                  username:
                   type: "string"
                   example: "Harshith"
                  email:
                   type: "string"
                   example: "example.p@gmail.com"
                  date_of_birth:
                   type: "string"
                   example: "2/1/1998"
                  assigned_to:
                   type: "string"
                   example: "details"
                  date_of_Joining:
                   type: "string"
                   example: "1/7/2021"
                  payment_terms:
                   type: "string"
                   example: "Terms"
                  role:
                   type: "string"
                   example: "Candidate"
                  isActive:
                   type: "boolean"
                   example: true
        "404":
          description: Failed request
          content:
           application/json:
            schema:
              type: "object"
              properties:
               status:
                type: "object"
                properties:
                 success:
                  type: "boolean"
                  example: false
                 code:
                  type: "integer"
                  example: 404
                 message:
                   type: "string"
                   example: candidate not found  
        "500":
          description: Failed request
          content:
           application/json:
            schema:
              type: "object"
              properties:
               status:
                type: "object"
                properties:
                 success:
                  type: "boolean"
                  example: false
                 code:
                  type: "integer"
                  example: 500
                 message:
                   type: "string"
                   example: error message   
          
    delete:
      tags:
      - "candidate"
      description: "Delete Candidate"
      parameters:
      - name: candidate_id
        required: true
        in: path
        schema:
          type: integer
      responses:
        "200":
          description: Success Request
          content:
           application/json:
            schema:
              type: "object"
              properties:
               status:
                type: "object"
                properties:
                 success:
                  type: "boolean"
                  example: true
                 code:
                  type: "integer"
                  example: 204
                 message:
                   type: "string"
                   example: request successful candidate deleted successfully
        "404":
          description: Failed request
          content:
           application/json:
            schema:
              type: "object"
              properties:
               status:
                type: "object"
                properties:
                 success:
                  type: "boolean"
                  example: false
                 code:
                  type: "integer"
                  example: 404
                 message:
                   type: "string"
                   example: candidate not found
        "500":
          description: Failed request
          content:
           application/json:
            schema:
              type: "object"
              properties:
               status:
                type: "object"
                properties:
                 success:
                  type: "boolean"
                  example: false
                 code:
                  type: "integer"
                  example: 500
                 message:
                   type: "string"
                   example: error message
 
components:
  schemas:
    Candidate:
     type: "object"
     required:
     - "name"
     - "email"
     - "date_of_birth"
     - "assigned_to"
     - "date_of_Joining"
     - "payment_terms"
     properties:
      name:
        type: "string"
        example: "Harshith"
      email:
        type: "string"
        example: "example.p@gmail.com"
      date_of_birth:
        type: "string"
        example: "2/1/1998"
      assigned_to:
        type: "string"
        example: "details"
      date_of_Joining:
        type: "string"
        example: "1/7/2021"
      payment_terms:
        type: "string"
        example: "Terms"